3:27 9/3/2025

+++ ThanTai/ThanTai/Models/DatHangChiTiet.cs
@@ -6,9 +6,9 @@
         public int DatHangID { get; set; }
         public int SanPhamID { get; set; }
         public short SoLuong { get; set; }
-        public int DonGia { get; set; }
-
-        public int TongTien { get;set; }
+        public decimal DonGia { get; set; }
+        public decimal TongTien { get;set; }
+        
         public DatHang? DatHang { get; set; }
         public SanPham? SanPham { get; set; }
     }
